[fixture] add style js global file
