+++ app/package.json
@@ -7,7 +7,8 @@
     "ios": "react-native run-ios",
     "lint": "eslint .",
     "start": "react-native start",
-    "test": "jest"
+    "test": "jest",
+    "postinstall": "patch-package"
   },
   "dependencies": {
     "@react-native-community/art": "^1.2.0",
@@ -37,6 +38,7 @@
     "eslint-plugin-react": "^7.32.2",
     "jest": "^29.2.1",
     "metro-react-native-babel-preset": "0.73.9",
+    "patch-package": "^6.5.1",
     "prettier": "^2.8.7",
     "react-test-renderer": "18.2.0",
     "typescript": "4.8.4"
