+++ app/src/components/Display/index.tsx
@@ -1,7 +1,8 @@
-import { StyleSheet, Text, View } from 'react-native';
+import { Text, View } from 'react-native';
 import { Neomorph } from 'react-native-neomorph-shadows';
 
 import React from 'react';
+import { DisplayStyles } from '../../../pages/calculator/style';
 
 interface DisplayProps {
     values: string[];
@@ -13,37 +14,18 @@ const Display = ({ values }: DisplayProps) => {
         const currencyValue = (parseFloat(number) / 100)
             .toFixed(2)
             .replace(/\B(?=(\d{3})+(?!\d))/g, ',');
-        return currencyValue == 'NaN' ? "0.00" : currencyValue;
+        return currencyValue == 'NaN' ? '0.00' : currencyValue;
     };
 
     return (
-        <View style={styles.view}>
-            <Neomorph inner swapShadows style={styles.neomorph}>
-                <Text style={styles.text}>R$ {toCurrencyValue(values)}</Text>
+        <View style={DisplayStyles.view}>
+            <Neomorph inner swapShadows style={DisplayStyles.neomorph}>
+                <Text style={DisplayStyles.text}>
+                    R$ {toCurrencyValue(values)}
+                </Text>
             </Neomorph>
         </View>
     );
 };
-
-const styles = StyleSheet.create({
-    view: {
-        alignItems: 'center'
-    },
-    neomorph: {
-        shadowRadius: 10,
-        borderRadius: 40,
-        backgroundColor: '#DDDDDD',
-        width: 330,
-        height: 125,
-        margin: '5%',
-        justifyContent: 'center'
-    },
-    text: {
-        color: '#818296',
-        alignSelf: 'flex-end',
-        fontSize: 32,
-        marginEnd: '5%'
-    }
-});
 
 export default Display;
