+++ app/src/components/Footer/index.tsx
@@ -1,22 +1,16 @@
 import { Neomorph } from 'react-native-neomorph-shadows';
-import {
-    TouchableOpacity,
-    Text,
-    StyleSheet,
-    BackHandler,
-    View
-} from 'react-native';
-
+import { TouchableOpacity, Text, View } from 'react-native';
 import Icon from 'react-native-vector-icons/FontAwesome5';
+import { FooterStyles } from '../../../pages/calculator/style';
 
 const Footer = () => {
     return (
-        <View style={styles.view}>
-            <Neomorph inner swapShadows style={styles.neomorph}>
+        <View style={FooterStyles.view}>
+            <Neomorph inner swapShadows style={FooterStyles.neomorph}>
                 <TouchableOpacity
                 // onPressIn={() => BackHandler.exitApp()}
                 >
-                    <Text style={styles.text}>
+                    <Text style={FooterStyles.text}>
                         <Icon name="calculator" size={20} color="#818296" />
                     </Text>
                 </TouchableOpacity>
@@ -25,23 +19,4 @@ const Footer = () => {
     );
 };
 
-const styles = StyleSheet.create({
-    view: {
-        paddingTop: '3%',
-        paddingBottom: '2%',
-        alignItems: 'center'
-    },
-    text: {
-        textAlign: 'center'
-    },
-    neomorph: {
-        shadowRadius: 3,
-        borderRadius: 10,
-        backgroundColor: '#DDDDDD',
-        width: 60,
-        height: 60,
-        // textAlign: 'center',
-        justifyContent: 'center'
-    }
-});
 export default Footer;
