--- app/style.js
@@ -1,27 +1,0 @@
-import { StyleSheet } from 'react-native';
-
-export default StyleSheet.create({
-    sectionContainer: {
-      marginTop: 32,
-      paddingHorizontal: 24,
-    },
-    sectionTitle: {
-      fontSize: 24,
-      fontWeight: '600',
-    },
-    sectionDescription: {
-      marginTop: 8,
-      fontSize: 18,
-      fontWeight: '400',
-    },
-    highlight: {
-      fontWeight: '700',
-    },
-    button:{
-      shadowRadius: 12,
-      borderRadius: 70,
-      backgroundColor: '#ECF0F3',
-      width: 140,
-      height: 140,
-    }
-  });
